ENGBL-736: Move select() call into specific recv_wrapper functions
We don't want to do the select call for UDTs as they may not pass in
a valid socket. Simply move the existing select call down a layer into
the specific transports for now. Also allow negative socket numbers
through as a UDT may pass one (even though it isn't using it).

Change-Id: I0d30d09b9b8bd0c93be2a98bfd60c098656090a0
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>