MAINT: Move files into new directory structure
Move the source files into a new directory structure.

/src - Source files
/src/transport - Source files for the transport layer

/test - Tests

/include/cmsg - Public header files

Note that in the future some of the header files should be made private
(cmsg_private.h, cmsg_queue.h etc). Also note that it is /include/cmsg
to match the include path used in the autogenerated files (<cmsg/cmsg.h>).

Change-Id: I0454643c7967c4991de7f3e18ae58ae4412c6256
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>