CR-58558: Renumber CMSG transport types
A previous patch: commit ad08c02 ("MAINT: Remove CMSG_TRANSPORT_RPC_LOCAL transport type")
removed the CMSG_TRANSPORT_RPC_LOCAL transport type. This inadvertently decremented
all of the enumeration values by 1. When a release using CMSG which still has the
CMSG_TRANSPORT_RPC_LOCAL type (i.e. pre 5.4.7-2) subscribes to a release with the
type removed (5.4.7-2 and later) the newer release will print an error that the subscriber
is trying to use an unsupported transport type.

This is because the older version is subscribing with what it thinks is CMSG_TRANSPORT_ONEWAY_TIPC=4,
but on newer releases 4 is now CMSG_TRANSPORT_CPG. To solve this simply for now
move CMSG_TRANSPORT_LOOPBACK to be the first enum value so that the values between
releases still match up.

Change-Id: I712372de8b54740a9daa75fcdddda6c2dc83dcec
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>