BLDSYS-206: Update CMSG compiler to protoc-c version 1.3.0
The CMSG compiler is based off the source code from the protoc-c compiler
from protobuf-c. Update the CMSG compiler with any changes that have
occurred between protoc v1.2.1 and v1.3.0.

Note that main.cc is not updated for now. This is because they have changed
the protoc-c compiler to be a plugin of the protoc (protobuf) compiler
rather than a standalone application.

Change-Id: I92366f0be79e59140cd9ebb3f3f6d85b8924094b
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>