CR-58784: Prevent accessing freed CMSG client from publisher
Repeated IMISH login and logout shows that memory corruption happens in CMSG
publisher code and causes ATMFD crashes with SEGV.

The current implementation in CMSG publisher uses the main thread to handle
subscription and notification messages are put in a queue. And other thread
is in charge of fetching messages from the queue and sending to the subscriber.

The problem happens while the notification send thread is on the way sending
a message when the subscriber (IMISH) exits, which causes send failure and
the send thread retries.
Meantime the main thread handles unsubscribe message and destroys the client
and transport. And when the send thread retries to send message it accesses
the memory pointed by client/transport (but already freed) and SEGV happens.

The issue is fixed in CMSG publisher library functions to defer destroying
client/transport if they are in use by messages:

1) There exists a logic that before deleting a subscriber all messages are
   cleaned up from the queue. But the problem is the notification send thread
   pops the message from the queue so still that message has a pointer to
   CMSG client that has been freed by the main thread.
   Now the notification send thread sets "in-use" flag for the subscriber entry
   that corresponding to the message being sent. This subscriber then is marked
   as "to-be-removed" instead of freeing immediately.

2) The "in-use" flag can be set/unset from both threads, so changed it from
   boolean to int.

3) When a subscriber entry is removed, make sure all messages are removed
   from the queue.

4) When "in-use" flag is set in a thread, make sure to clean up the subscriber
   entry if it is marked as "to-be-removed" after unset "in-use" flag.

5) cmsg_pub_subscriber_remove_all_with_transport()
   Freeing messages are done in _cmsg_pub_subscriber_delete_link() now so
   do not need to call cmsg_send_queue_free_all_by_transport() explicitly.

6) cmsg_pub_invoke()
   Make sure clean up the subscriber if marked as "to-be-removed" by calling
   _cmsg_pub_subscriber_cleanup_link().

7) Mutex order
   Make the mutex lock happens in the following order to avoid self-lock:
    - publisher->subscriber_list_mutex
    - publisher->queue_mutex

8) _cmsg_pub_queue_process_all_direct()
   Set the subscriber's "in-use" flag with proper locks

Change-Id: Ibcc3b6c35dac3008ef730e6cf3674f4423e799a7
Reviewed-by: Justin Jose <justin.jose@alliedtelesis.co.nz>
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>