VISTA548-61: Fix broken build
The build got broken with the last patch. The cmsg_proxy and
cmsg_proxy_passthrough functions had an extra parameter added but the unit
tests weren't updated to match, so test32 target broke.

This patch updates the unit tests and adds a description of the new
parameter to the comment before the two functions.

Change-Id: I34680693e228d07c590195c76b8311fa659e94c6
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>