ENGBL-917: Don't generate #includes for http.proto dependencies
If a proto file imports http.proto it is only doing it so that the
protoc/protoc-cmsg compilers can correctly parse the 'http' option. We
don't run the protoc-cmsg compiler over http.proto so we don't want to
generate #includes in dependent files for auto-generated http files
(as they wont exist).

Change-Id: I727cc54b91dbd168515ef77d155409588c85d7e8
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>