GUIAPI548-95: Convert cmsg-proxy to use structure for input/output
The function was beginning to use a large number of input and output
parameters, so these have been wrapped into structures. This will also
avoid having to change the function prototype in the future.

Freeing the contents of the output structure has been made into an API
call. This wraps the free and cmsg_proxy_free_extra_headers that were
previously required.

All the unit tests are updated and passing.

Change-Id: Ibbfedfced482e86cb360e56410784557dc2371de
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>