ENGBL-647: Add broadcast client/server application framework
This is initially intended as a replacement for the TIPC broadcast transport
and the TIPC broadcasting implemented using a composite client and stack events
in many places in AW+. Some form of arbitration would be required to be added
to support replacing CPG with this (as CPG currently gives guaranteed ordering
of delivery).

A couple of functional tests are also implemented to prove the functionality
works as intended.

Change-Id: Ib6596710369a9c46ce39485163252a60668692aa
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>