ENGBL-931: Server side validation (1): Add framework and test cases
Validating messages can produce a lot of boilerplate. This patch series
intends to add functionality to the cmsg compiler allowing it to generate
message validation functions using an extension.

Furthermore, for methods that return an 'ant_result' message we wish for
this new validation function to be called automatically and respond
accordingly if the message does not validate. This further functionality
will be implemented using yet another related extension.

This first patch simply adds the extension framework and adds test cases
showing how this functionality is expected to work. Note that the tests
do not currently compile as the autogenerated validation function is not
yet being generated.

Change-Id: I15a41e5e01850be794408571b32081a11d1be39e
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>