CR-58926: Fix cmsg_pub_queue_free_all if queue is re-used
Prior to this fix, cmsg_pub_queue_free_all() set the variable
queue_process_count to -1 when it freed the queue. Zero would seem a
more-correct value, however this wouldn't stop
cmsg_pub_queue_process_all() from waiting forever for the count to
become non-zero. However, setting it to -1 causes issues if the queue is
used again (which it is in poehw). Firstly, the count is unsigned, so -1
is a big number. Also, adding something to the queue could potentially
make this zero.

This patch changes the API so that the thread that we were trying to
stop is now handled within the cmsg library itself. The thread is
created by calling the new function cmsg_pub_queue_thread_start(), and
is terminated by cmsg_pub_queue_thread_stop(). It will also be
terminated when the publisher is destroyed. The function
cmsg_pub_queue_free_all() now only frees the queue, and does not try to
signal cmsg_pub_queue_process_all().

Change-Id: I68cf87c2596889daec9842fb906feb76da1dd6a8
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>