ENGBL-647: Handle broadcast server listening in another thread
By abstracting it inside the library and doing it in a separate
thread it should not be possible for connection attempts to fail
due to timeouts (assuming there is nothing wrong with the OS).

Also add in some pthread_setcancelstate calls so we are explicit where
the separate threads will be cancelled.

Change-Id: I6533ff0742d19c44810f079cd1e67a5b2c33452f
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>