GUIAPI548-124: Use bytes type for special _file field
This is part of a change to use the protobuf 'bytes' type for the
hidden _file field, instead of a repeated uint8 field. This is a
more intuitive choice for this use case, and means that we don't
need to modify third-party software such as protoc-gen-swagger to
get the documentation output that we want.

Setting the value of a bytes field is slightly inconsistent with
setting other types of fields (it is a pointer inside a struct)
so a new macro has been added to cmsg.h for setting a field of
bytes type.

The cmsg-proxy functional tests have been updated accordingly.

Change-Id: I3a7b08d341cab590077db16684702e5eb5fe7a97
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>