ENGBL-647: Pass accepted socket by pointer to integer
If stdin and stdout are closed then 0 is a valid socket descriptor that
could be returned. This will be cast to NULL meaning it would be impossible
to tell whether the pointer is NULL, or a value of zero cast to NULL. Simply
allocate an integer to pass on the queue instead.

Change-Id: I62c09376c06fb2c3456e205a49aa7b74fbc6f4ff
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>