ENGBL-721: Ensure ProtobufCServiceDescriptor is always const
With the arc compiler, if the -wcast-qual cflag is used, certain
casts are rejected if they discard the const attribute.

The CMSG_DESCRIPTOR macro casting the const descriptor to a non-const is
one of these. As these descriptors are always const, this patch updates
the code so that the cast is not needed.

Change-Id: I06e81d22ae555dfbab19ce697ea2aa6e81c89dc5
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>