VISTA548-123: Strip streamed response JSON similarly to normal proxy output
We have some existing logic used for normal cmsg proxy responses where
if there is only a single field in the message (ignoring the included
ant_result sub message) then we strip the key off and just return the
data. Do the same for the stream responses so that they match the format
shown as output in the documentation.

Change-Id: Ifc68bbd1fe3a55e84b07aa37d3fa607f50033989
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>