CMSGPROX-181: Add configuration library for CMSG proxy
This library stores the configuration for CMSG proxy (for now only
the logging mode). The configuration can be set by any process that
links the library. It is implemented in shared memory because the
proxy may not always be up and running to receive the configuration
request (the HTTP server using the proxy can be stopped and restarted).

Change-Id: I884bf3a38fb52e907f4ca536f262600792b3576b
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>