CMSGPROX-202: Add some blank comments at the top of ant_result.proto
When the proto files are copied to the staging directory, they are passed
through the C preprocessor. The cpp arguments are set to preserve comments,
which unfortunately also preserves some CPP comments.

https://stackoverflow.com/questions/25314725/remove-the-comments-generated-by-cpp

Because blank lines are removed, the last comment was being treated as the
description of the ant_code enum and appearing in the documentation.
("We do not support C11 <threads.h>")

By adding some blank comments at the top, we avoid this problem. This
doesn't affect other files because they generally have import statements
at the top.

N.B. The lines with 3 slashes are stripped prior to docs generation.

Change-Id: Ib4b53b213748f4265491f6c43a50b4a6cb4372a1
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>