GUIAPI548-322: Input URL is now encoded
Currently the URL input to cmsg-proxy is decoded. This exposes a
security flaw (CWE-117: Improper Output Neutralization for Logs)
when we log the URI that was accessed. By passing in the URI undecoded
we can instead log the raw URI, removing the security flaw.

Simply update the code to ensure each URL parameter is decoded as it
is parsed from the URL string.

Change-Id: Ic36ee9542cd81319195311eacd70d762b32da107
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>