VISTA548-289: Prevent using CMSG_IS_PTR_PRESENT() with integer member
Sometimes it's to make a mistake when checking a field is present.

By forcing to compare with NULL "(void *) 0)", gcc will generate a compiler
warning "comparison between pointer and integer" if CMSG_IS_PTR_PRESENT()
is used instead of using CMSG_IS_FIELD_PRESENT().

The opposite is already detected because "has_*" field does not exist
for pointer type member.

Change-Id: I297f840760e855c969ad75ba918dd1dd4eca0c1e
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>