VISTA548-306: Cancel streaming if conversion to JSON fails
Currently if the JSON to protobuf conversion fails then the request
hangs as output->stream_response is true meaning appweb expects the
response to be streamed but cmsg proxy has finished processing.

Simply ensure we cancel the streaming if the conversion fails. Note
that we move the streaming connection creation after grabbing the
client to avoid more cleanup code if finding the client fails.

Change-Id: I1e7be47cce759950544f4ab0c8fff28dec3edc9d
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>