CR-60194: Reimplement the loopback client/transport
The current implementation of the loopback client/transport uses a
pipe internally. However, the pipe can only have so much data written
onto it before it is full. Since we are invoking the server directly
it is not possible to read any data off the pipe until the entire message
is written. If the message is too large then the write to the pipe will
fail and only a partial message will be sent.

To solve this limitation it is simplest to just reimplement the loopback.
Instead of writing the message onto a pipe just copy the message into
dynamic memory and read this data as required in the client receive.

Note that we now use the user-defined transport to implement the loopback
transport. This is simply to reduce code duplication and we do use the
user defined data field to store the copied message.

Change-Id: I39f1dc5266ce0bb127a70e330b3b5bf242d5c062
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>