ENGBL-1045: Add shell application for calling CMSG APIs
Create a simple application that can be used from the shell to call
CMSG APIs. Initially only unix and tipc transports are supported (TCP
could be added if required in the future). Note that the input message
is expected in JSON format and any received message is printed in JSON
format (with pretty printing).

Change-Id: I738c7e327dafcec53c42a5f2aa218f21fcda5e4a
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>