ENGBL-949: Add pthread helpers for unix publishers/subscribers
Add more pthread helpers that can be used to start a publisher or
subscriber in a new pthread. A functional test is added to prove
the helpers work. Note that the test needs to unsubscribe to avoid
a socket leak found by valgrind. The unsubscribe path cleans up the
subscriber the same as 'cmsg_pub_subscriber_remove_all' so the cause
of this supposed leak is unclear.

Change-Id: I8fbaa2e70602a122807ef16f6cbe69da38f1af0b
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>