VISTA548-346: Add string format validation and custom error messages to validation
Update the validation extension to automatically generate validation
for string fields. Use an enum to specify the specific type of validation
to do. Note that specifying regex was also considered but this seems overly
complicated for common formats such as an IP address.

Add functional tests to prove the functionality behaves as we expect.

As can be seen this patch completely re-writes the validation logic to achieve
the required change. We now simplify the implementation by defining the required
validation functions inside the cmsg library and simply call into these in the
generated code.

Change-Id: Ic52f0e0104cade3a811e9e678e37df90c071ad12
Reviewed-by: Bernie Harris <bernie.harris@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>