ENGBL-1113: Don't make accept thread noncancellable before accept call
When the thread is made noncancellable it cannot be cancelled until it
is made cancellable again. Even though the cancellation request is
received when a cancellation function is called (select, accept etc) it
will not be processed until the thread is made cancellable. This is a
problem with the accept call as it will never complete until a connection
is made. Therefore the thread is not being cancelled when requested and
the pthread_join in cmsg_server_accept_thread_deinit is blocking indefinitely.

The intent of making the thread noncancellable is to ensure that an accepted
socket is not leaked on the small chance the thread is cancelled right as it is
accepted. Only make the thread noncancellable as we add the socket onto the queue
as this is the only potential cancellation point that would cause the socket to
be leaked. From the pthread documentation I don't think the calloc call is a
cancellation point, but I've put it inside the noncancellable part anyway.

Change-Id: Ia280940ddf2ebeec143f7086fd7a1f43f7917d35
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>