DFIPL-30: Add new function to suppress error message from CMSG server
Similar to CMSG client, new API cmsg_server_suppress_error() is introduced.
This allows applications can suppress error-level syslog from CMSG server.
If it's turned on error-lever log is generated as debug-level log.

This flag is also passed down to the transport. This change is made to CMSG
server and client.

Change-Id: Ia5d5ec3983851c605ab4b2f35d09d8869048e026
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>