CR-60952: Don't access if a message field is NULL
When the auto validation code is genereated, the field existence
is not checked before accessing it. If date specific summer time is configured,
recurring config message is not given and imi crashes when the NULL recurring
message field is accessed.

Auto code generation is changed to check the existence of the message field
before accessing it.

Change-Id: I43c6645400066014e521f5d62e52f6c32f543bd7
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>