CR-61071: Don't allow cmsg-proxy to accept empty data if input is expected.
If we expect something to be set in the body, we do not want to allow
empty input. Also check to ensure that we don't add APIs where we say we
expect something in the body but don't actually expect any.

A check is added to ensure that data is present when expected, and the
opposite check to ensure that no data is present when not expected is
moved to the same function.  This was doing an additional check with
body string "*" that should be done at install time. A check and error
message has been added at install time for this case.
There was one existing case which will be fixed up.

* unnecessary error_msg variable is removed.
* minor warning fixup in cmsg_proxy_http_streaming_api.c

Also fixup existing unit tests that don't meet these restrictions and add
tests for new functionality. (includes some fix-indent).

Change-Id: Idf245269f90364306646af44ec019abd965b46a8
Reviewed-by: Bernie Harris <bernie.harris@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>