CR-61108: Ensure streaming connection lookup returns correct connection
Currently if a connection has been removed then
cmsg_proxy_streaming_lookup_conn_by_id will return the last connection
iterated through in the for loop. This causes data from multiple closed
streams to start being written to another connection that is still open.
Fix the lookup to ensure it returns NULL if the connection has been deleted.

Also take the opportunity to use a common style in the connection lookups.
Don't use a bool to specify if it was found, just perform the action in place
and break/return early.

Change-Id: Ibcb0a3070e7fcb73bceb8170e1ae76e647c5905f
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>