ENGBL-1184: Use single connection accepting thread
To reduce number of threads inside programs using
CMSG servers, change cmsg_server_accept_thread_init
to use a singular global thread.

Now epoll is used instead of select. There is no
nice way of adding new FDs to be polled in one
thread while another thread is blocking on select.
With epoll, adding a new FD with epoll_ctl in one
thread while epoll_wait is running in another is
valid.

The thread field on the info is now unused and removed.
Also free the info if the init fails as this memory
was being lost if this function failed.

Change-Id: Ie2d2a06f3d8bec6597a60aa3b351427862db68aa
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>