CMSGPROX-38: Ensure a message is always returned with ANT_CODE_INTERNAL.
This patch ensures that when ANT_CODE_INTERNAL is produced, and
output->http_status is set to HTTP_CODE_INTERNAL_SERVER_ERROR, it is done
with a call to cmsg_proxy_generate_ant_result_error. It also removes null
error messages and makes sure that a descriptive error message is always
set.

Change-Id: I869e13faf4c0c809793f31fd82d060ed639b73b5
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>