GUIAPI548-463: Create new client/server create functions for TCP
The existing functions for creating a TCP client or server are convoluted
in that they take a cmsg_socket structure to setup. Create new functions
similar to the TIPC ones in that they just take the service name (to find
the port number in the /etc/services file) and the IP address to use. The
one user of the old functions can be updated to use the new ones in the
future.

Change-Id: Ifc1ef2b377b1f0e20b0b360fdd3e9e2e91642a85
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>