GUIAPI548-491: Transport now only contains a single socket
Previously there was two sockets on the transport which was confusing.
On clients the 'client_socket' was the only socket in use. However on
servers the 'listening_socket' was what the server listened on for
connections, while the 'client_socket' held the socket of the current
connection the server was processing (so that it could send the response).

Now the current connection is no longer stored in 'client_socket' meaning
we can reduce the number of sockets to one.

Change-Id: I8dafff78d10ee3ca691f4323352aa66c7da9af52
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>