GUIAPI548-491: Create pthread helper to run a server in multi-threaded mode
Now that the CMSG server functionality is thread safe it is possible to
run multiple connections each in a separate thread. Previously each
connection would have to be processed serially. Add the required code
to run the server in such a way and provide a simple API for users of
the CMSG library to use this functionality.

Note that to avoid potential excessive thread usage we explicitly close
connections (and subsequently exit the thread) if they have not had any
activity in the last 5 minutes. This behaviour may not actually be
required, but it seems like a safe sanity check to avoid unintentional
resource usage.

Change-Id: Ia1d1f4a3b3905fb0ae5d5fb4a4df1795998c5370
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>