GUIAPI548-458: Add support for streaming raw file data
This is part of a change that adds a web API for downloading
files from a device's file system.

Add a new RPC for sending raw file data to be streamed. Data
passed to this RPC does not need to be packed. Also record the
length of the data because strlen can't be used with raw file
data.

Add the additional newline to plaintext or json output here
instead of in cmsgProxyHandler.

Sending the correct headers with the data will be handled in a
separate patch.

Change-Id: Idf0f92ebaea4ee5cec70aeb843c5d3cb89d73cca
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>