GUIAPI548-458: Add RPC to streaming API for setting response headers
For a streaming response, the HTTP headers should be set from the
same thread that sends the response data in blocks. This patch
adds an RPC to the HTTP streaming API for setting the correct
headers for APIs that stream data to the caller. If the content
type is raw file data, the name of the file being transferred as
well as the file size is passed to the set_headers RPC, which
uses the information to correctly set the 'Content-Disposition'
and 'Content-Length' header values.

Also add helper functions for freeing cmsg_proxy_stream_response_data
and cmsg_proxy_stream_header_data structs.

Remove the cmsg_proxy_generate_file_response function because it
is no longer needed. Raw file data will be sent to the caller via
a streamed response.

Rename the header key/value variables since these being declared
in multiple places causes building the tests to fail.

Also, remove the tests involving sending file data in the
response. These will be rewritten in a later patch since
responses containing file data are now always streamed.

Change-Id: Ie966b686956d6e3336ff49fec50f52272ef32c94
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>