GUIAPI548-458: Ensure headers are set before streaming response
This patch adds a flag to the cmsg_proxy_stream_connection struct
that is used to check whether the headers for a streaming
connection have been set yet. The flag is set in the
set_stream_headers RPC, and the send_stream_data and
send_stream_file_data RPCs will log an error if the flag is not
set yet.

Also, don't allow sending a response using send_stream_data if
the output message has a '_file' field, or using
send_stream_file_data if the output message does not have a
'_file' field.

Change-Id: Ib2bd98f98a647b45c9a0b8a2438b85fcb47c4231
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>