GUIAPI548-458: Avoid null-reference by multi-threaded server
Currently, before a server executes an impl function, it sets the
'closure' and 'closure_data' fields within the service structure
and then afterwards sets them to NULL. In a multi-threading
situation this can result in a NULL reference if one thread of
the multi-threaded server sets the fields to NULL before another
thread has finished executing.

This patch fixes this problem by passing the closure information
using a locally declared struct, instead of using the service
structure which is shared by all threads.

The functional tests have been updated to account for this
change, including the multi-threading test which has been
modified so that without this fix the error condition will occur.

Change-Id: Idfa7fc0bb0b852d805ca221a294592aaf4051a15
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>