GUIAPI548-458: Correctly close all sockets on the streaming server
The fdmax parameter of cmsg_server_receive_poll() expects the
value of the maximum fd number, not max + 1 (as is expected by
a call to select()) When a new socket is opened, the maximum file
descriptor value is assigned to fdmax. Because of this,
cmsg_proxy_streaming_server_cleanup() should loop through the
file descriptor values up to and including fd_max.

The function comment for cmsg_server_receive_poll() has been
updated to clarify its usage.

A similar fix is applied to pthread_server_run() and
pthread_server_cancelled().

Additionally, the test_http_streaming and test_http_file_streaming
functional tests have been updated to wait until all the required
stages of the test have finished before exiting, instead of just
sleeping for a long enough period of time. The changes to the tests
depend on the fix implemented in this patch.

Change-Id: Ifeb820029cd50341239ca2f2014566c7e4bbd1a6
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>