GUIAPI548-492: Modify cmsg_proxy_input struct to allow for file upload
This patch adds the cmsg_proxy_api_file_info structure to the
cmsg_proxy_input structure. This is necessary for the file upload API.
It allows for information about the file uploaded with appweb to be passed
through to filesysd. cmsg_proxy_api_file_info contains the temporary
filename set by appweb during the upload and the original filename given by
the user. These fields are set as internal api values in
cmsg_proxy_input.c.

Change-Id: I6197e8f884bb2c73ffa51b14662e9630248041d0
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Bernie Harris <bernie.harris@alliedtelesis.co.nz>
Signed-off-by: ’Laura <laura.king@alliedtelesis.co.nz>