GUIAPI548-462: Add RPC to streaming API for aborting connection
This patch is part of a change to add fastfailover support to
filesysd, so that nodes leaving during a download operation can
be handled properly.

An RPC has been added to the streaming API which forces an
existing connection to abort. The abort_stream_connection API
schedules a call to httpError() in appweb.

Additionally, some minor details have been cleaned up in the
streaming code: wrong/missing comments, unused variables, and
replacing g_list_remove with g_list_delete_link which is more
efficient, and using the proper Send functions from impls.

Change-Id: Ic155836c274e98d2079beddc6b755ed9f0ae20aa
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>