CR-62843: Add check to TLV processing to prevent segfault
CMSG TLV processing could come across a TLV value length
that is larger than the expected size of the remaining
headers.  This would cause an integer overflow (by
subtracting a larger number from a smaller number) and
cause invalid memory to be accessed.

Add a check for the potential overrun and prevent it
by returning an error to stop processing the message.

Change-Id: I4973b2808e475b1429f7bfc2db79140b50ed6572
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>