ENGBL-1305: Notify remote hosts when servers are created/destroyed
Add the required code to ensure that when servers are created/destroyed
on the local host we sync this information to remote hosts. Note that we
only need to sync servers that use TCP and the same IP address as that
of the remote sync server (as these are the only servers that the remote
host will be able to connect to).

Change-Id: I563378d6fa53c369c02e881326898c18cec30f2d
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>