ENGBL-1306: Store subscriptions in cmsg_pssd
To begin the rework of the publisher subscriber functionality store
all subscriptions in the new cmsg_pssd application. Implement the
debug functionality to print the stored subscriptions when SIGUSR1
is received by cmsg_pssd.

Note that to avoid chances of regressions we choose to implement support
for TIPC transports for now.

Change-Id: I06994efd485b372f57a520be2a6e41ac3bd4c2f1
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>