ENGBL-1305: Improve usability of service listener functionality
Currently we provide a CMSG server that the user of the service listener
functionality can poll to process any service listener events. This is not
the most usable solution as it forces the user to do all of their processing
from a single thread. Furthermore it would not be possible for two separate
threads to listen for the same service. Therefore we rewrite this functionality
to be implemented using an eventfd/queue combo to maximise the usability and
also closer match the TIPC topology service functionality.

Change-Id: Icfce073cb4e72c0ec90c5c21d9b3f6dd155e2f86
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>